Bugs de "TipoServicio" corregidos, actualizar implementado y estilos base agregados al 100%.

+++ src/java/Interfaces/iTipoServicioDAO.java
@@ -4,9 +4,9 @@ import Modelo.TipoServicio;
 import java.util.List;
 
 public interface iTipoServicioDAO {
-    public byte insertar(TipoServicio ts);
+    public int insertar(TipoServicio ts);
     public List<TipoServicio> buscar(String nts);
     public TipoServicio obtenerPorId(int id);
-    public byte actualizar(TipoServicio ts);
-    public byte eliminar(int id);
+    public int actualizar(TipoServicio ts);
+    public int eliminar(int id);
 }
